Detect when within multi line-comment block
